Epmlstrtai 113

+++ Java/JDI/jdi-cucumber-test-generator/src/main/java/com/epam/test_generator/controllers/SuitController.java
@@ -20,39 +20,78 @@ public class SuitController {
     SuitService suitService;
 
 	@RequestMapping(value = "/")
-    public String getMainPage(){
+    public String getMainPage() {
         return "/WEB-INF/static/views/newSuits";
     }
 
     @RequestMapping(value = "/suits", method = RequestMethod.GET, produces = "application/json")
     @ResponseBody
     public ResponseEntity<List<SuitDTO>> getSuits() {
-
-        return new ResponseEntity<>(suitService.getSuits(),HttpStatus.OK);
+		return new ResponseEntity<>(suitService.getSuits(), HttpStatus.OK);
     }
 
     @RequestMapping(value = "/suit/{suitId}", method = RequestMethod.GET)
     @ResponseBody
     public ResponseEntity<SuitDTO> getSuit(@PathVariable("suitId") long id){
+        SuitDTO suitDTO = suitService.getSuit(id);
 
-        return new ResponseEntity<>(suitService.getSuit(id), HttpStatus.OK);
+		if (suitDTO != null) {
+			return new ResponseEntity<>(suitDTO, HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @RequestMapping(value="/suit/{suitId}", method = RequestMethod.PUT, consumes = "application/json")
-    public ResponseEntity<Void> editSuit( @PathVariable("suitId") long id, @RequestBody SuitDTO suit){
-        suitService.updateSuit(suit);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<Void> editSuit( @PathVariable("suitId") long id, @RequestBody SuitDTO suitDTO){
+        if (isNameValid(suitDTO.getName()) && isPriorityValid(suitDTO.getPriority()) &&
+                isDescriptionValid(suitDTO.getDescription()) && isTagsValid(suitDTO.getTags())) {
+            suitService.updateSuit(suitDTO);
+
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(HttpStatus.UNPROCESSABLE_ENTITY);
     }
 
     @RequestMapping(value = "/suit/{suitId}", method = RequestMethod.DELETE)
     public ResponseEntity<Void> removeSuit(@PathVariable("suitId") long id){
-        suitService.removeSuit(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        SuitDTO suitDTO = suitService.getSuit(id);
+
+        if (suitDTO != null) {
+            suitService.removeSuit(id);
+
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @RequestMapping(value="/suit", method = RequestMethod.POST, consumes = "application/json", produces = "application/json")
     @ResponseBody
-    public ResponseEntity<SuitDTO> addSuit(@RequestBody SuitDTO suit) {
-        return new ResponseEntity<>(suitService.addSuit(suit), HttpStatus.OK);
+    public ResponseEntity<SuitDTO> addSuit(@RequestBody SuitDTO suitDTO) {
+        if (isNameValid(suitDTO.getName()) && isPriorityValid(suitDTO.getPriority()) &&
+                isDescriptionValid(suitDTO.getDescription()) && isTagsValid(suitDTO.getTags())) {
+			return new ResponseEntity<>(suitService.addSuit(suitDTO), HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(suitDTO, HttpStatus.UNPROCESSABLE_ENTITY);
     }
+
+    private boolean isPriorityValid(Integer priority) {
+        return priority != null && priority >= 1 && priority <= 5;
+    }
+
+    private boolean isNameValid(String name) {
+        return name != null && name.length() >= 1 && name.length() <= 255;
+    }
+
+    private boolean isDescriptionValid(String description) {
+        return description == null || description.length() <= 255;
+    }
+
+    private boolean isTagsValid(String tags) {
+        return tags == null || tags.length() <= 255;
+    }
+
 }
